+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Version counting is based on semantic versioning (Major.Feature.Patch)
 
 ### YACReader
 * Improve double page layout stability when going back and forth after finding single image exceptions.
+* Add more system information about the rendering pipelines used.
 
 ## 10.3.1